moved artboard.server to model/artboards dir
