+++ app/definitions/artboard-generator.ts
@@ -1,5 +1,5 @@
 import { type IArtboardVersion } from '#app/models/artboard-version/artboard-version.server'
-import { type IArtboard } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { type IFill } from '#app/models/fill.server'
 import { type ILayer } from '#app/models/layer.server'
 import { type ILayout } from '#app/models/layout.server'
