+++ app/models/design/design.create.server.ts
@@ -3,7 +3,7 @@ import { type designTypeEnum } from '#app/schema/design'
 import { NewArtboardVersionDesignSchema } from '#app/schema/design-artboard-version'
 import { prisma } from '#app/utils/db.server'
 import { type IArtboardVersion } from '../artboard-version/artboard-version.server'
-import { type IArtboard } from '../artboard.server'
+import { type IArtboard } from '../artboard/artboard.server'
 import { type IDesign } from '../design.server'
 import { type ILayer } from '../layer.server'
