+++ app/models/layer.server.ts
@@ -6,7 +6,7 @@ import {
 } from '#app/schema/layer'
 import { prisma } from '#app/utils/db.server'
 import { type IArtboardVersion } from './artboard-version/artboard-version.server'
-import { type IArtboard } from './artboard.server'
+import { type IArtboard } from './artboard/artboard.server'
 import { type IDesignWithType } from './design.server'
 
 export interface ILayer {
