+++ app/models/layer/layer.create.server.ts
@@ -4,7 +4,7 @@ import { ValidateArtboardVersionParentSubmissionStrategy } from '#app/strategies
 import { validateEntitySubmission } from '#app/utils/conform-utils'
 import { prisma } from '#app/utils/db.server'
 import { type IArtboardVersion } from '../artboard-version/artboard-version.server'
-import { type IArtboard } from '../artboard.server'
+import { type IArtboard } from '../artboard/artboard.server'
 import { type ILayer } from '../layer.server'
 import { type IUser } from '../user/user.server'
