+++ app/routes/sketch+/artboards+/$slug_+/components/canvas-content.tsx
@@ -14,7 +14,7 @@ import {
 	type IArtboardGenerator,
 } from '#app/definitions/artboard-generator'
 import 'reactflow/dist/style.css'
-import { type PickedArtboardType } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { canvasDrawService } from '#app/services/canvas/draw.service'
 
 const initialNodes = [
@@ -30,7 +30,7 @@ export const CanvasContent = ({
 	artboard,
 	artboardGenerator,
 }: {
-	artboard: PickedArtboardType
+	artboard: IArtboard
 	artboardGenerator: IArtboardGenerator | null
 }) => {
 	const { width, height, backgroundColor } = artboard
