+++ app/routes/sketch+/artboards+/$slug_+/components/forms/design/panel-form-design-new.tsx
@@ -3,7 +3,7 @@ import { getFieldsetConstraint } from '@conform-to/zod'
 import { useActionData, useFetcher } from '@remix-run/react'
 import { AuthenticityTokenInput } from 'remix-utils/csrf/react'
 import { PanelIconButton } from '#app/components/ui/panel-icon-button'
-import { type IArtboard } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { type ILayer } from '#app/models/layer.server'
 import {
 	type NewDesignSchemaType,
