+++ app/routes/sketch+/artboards+/$slug_+/components/panel/artboard/panel-artboard-layers.tsx
@@ -9,7 +9,7 @@ import {
 	SidebarPanelRowValuesContainer,
 } from '#app/components/templates'
 import { Icon } from '#app/components/ui/icon'
-import { type PickedArtboardType } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { type ILayer } from '#app/models/layer.server'
 import { PanelFormArtboardLayerNew } from '../../forms/artboard/layer/panel-form-artboard-layer-new'
 import { PanelFormArtboardLayerReorder } from '../../forms/artboard/layer/panel-form-artboard-layer-reorder'
@@ -20,7 +20,7 @@ export const PanelArtboardLayers = ({
 	artboard,
 	layers,
 }: {
-	artboard: PickedArtboardType
+	artboard: IArtboard
 	layers: ILayer[]
 }) => {
 	const [searchParams, setSearchParams] = useSearchParams()
