+++ app/routes/sketch+/artboards+/$slug_+/components/panel/design/sidebar-panel-header-design.tsx
@@ -2,7 +2,7 @@ import {
 	SidebarPanelHeader,
 	SidebarPanelRowActionsContainer,
 } from '#app/components/templates'
-import { type IArtboard } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { type ILayer } from '#app/models/layer.server'
 import {
 	type NewDesignSchemaType,
