+++ app/routes/sketch+/artboards+/$slug_+/components/panel/design/type/panel-design-type-row.tsx
@@ -2,7 +2,7 @@ import {
 	SidebarPanelRow,
 	SidebarPanelRowContainer,
 } from '#app/components/templates'
-import { type IArtboard } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { type IDesign, type IDesignIdOrNull } from '#app/models/design.server'
 import { type ILayer } from '#app/models/layer.server'
 import {
