+++ app/schema/entity.ts
@@ -1,6 +1,6 @@
 import { type IArtboardBranch } from '#app/models/artboard-branch/artboard-branch.server'
 import { type IArtboardVersion } from '#app/models/artboard-version/artboard-version.server'
-import { type IArtboard } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { type IDesignWithType, type IDesign } from '#app/models/design.server'
 import { type IFill } from '#app/models/fill.server'
 import { type ILayer } from '#app/models/layer.server'
