+++ app/services/artboard/design/toggle-visible.service.ts
@@ -1,5 +1,5 @@
 import { type User } from '@prisma/client'
-import { type IArtboard } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import { type IDesign, type IDesignIdOrNull } from '#app/models/design.server'
 import { ArtboardUpdateSelectedDesignStrategy } from '#app/strategies/design/update-selected.strategy'
 import { designToggleVisibleService } from '../../design/toggle-visible.service'
