+++ app/strategies/design/create.strategy.ts
@@ -1,6 +1,6 @@
 import { type User } from '@prisma/client'
 import { type IArtboardVersion } from '#app/models/artboard-version/artboard-version.server'
-import { type IArtboard } from '#app/models/artboard.server'
+import { type IArtboard } from '#app/models/artboard/artboard.server'
 import {
 	createArtboardVersionDesign,
 	createDesign,
